Allow multiple, configurable reporters
Start to try and unpick reporter so it can handle multiple outputs

Console reporter ~66% done

Refactor mostly there

Backwards Comaptibility

config "reporter" property to be deprecated in the future

Attempting to fix

Allow multiple reporters
