+++ lib/reporter/index.js
@@ -1,5 +1,6 @@
 var session = require('moonraker').session,
     config  = require('moonraker').config,
+    reporters = require('moonraker').reporters,
     Base    = require('mocha').reporters.Base,
     i18n    = require('./i18n'),
     fs      = require('fs'),
@@ -47,7 +48,8 @@ function Moonraker(runner) {
     if (suite.root) return;
     if (level === 0) {
       finishFeature(result, feature);
-      printFeature(feature);
+      // printFeature(feature);
+      reporters.emit('featureFinished', feature);
       feature = null;
     } else {
       feature.scenarios.push(scenario);
@@ -81,6 +83,8 @@ function Moonraker(runner) {
   });
 
   runner.on('fail', function (test) {
+    console.log('FAIL!');
+    console.log(test);
     var errorStr = test.err.stack || test.err.toString();
     if (!~errorStr.indexOf(test.err.message)) {
        errorStr = test.err.message + '\n' + errorStr;
@@ -93,7 +97,8 @@ function Moonraker(runner) {
       screenshotPath = test.title.replace(/\W+/g, '-').toLowerCase() + '.png';
     }
     if (!feature) {
-      console.log(errorStr);
+      reporters.emit('featureError', errorStr);
+      // console.log(errorStr);
     }
     if (scenario) {
       scenario.steps.push({
@@ -115,34 +120,4 @@ function finishFeature(result, feature) {
       for (var attr in feature) { f[attr] = feature[attr]; }
     }
   });
-}
-
-function printFeature(feature) {
-  if (config.icinga) return;
-  var buffer = '';
-  buffer += util.format('%s: %s (finished on thread: %s)\n', i18n.get('feature'), feature.title, feature.thread);
-
-  feature.scenarios.forEach(function (scenario) {
-    buffer += util.format('   %s: %s\n', i18n.get('scenario'), scenario.title);
-
-    scenario.steps.forEach(function (step) {
-      if (step.status == 'pass') {
-        buffer += util.format(color('checkmark', '     %s'), Base.symbols.ok);
-        if (step.speed == 'fast') {
-          buffer += util.format(color('pass', ' %s \n'), step.title);
-        } else {
-          buffer += util.format(color('pass', ' %s '), step.title);
-          buffer += util.format(color(step.speed, '(%dms)\n'), step.duration);
-        }
-      } else if (step.status == 'fail') {
-        buffer += util.format(color('fail', '     X %s\n'), step.title);
-        buffer += color('fail', step.error + '\n');
-      } else if (step.status == 'pending') {
-        buffer += util.format(color('pending', '     - %s\n'), step.title);
-      } else if (step.status === 'skipped') {
-        buffer += util.format(color('medium', '     - %s\n'), step.title);
-      }
-    });
-  });
-  console.log(buffer);
 }
